Fix and refactor JDBC ITs and ddl
