+++ addons/common/persistence/jdbc/src/main/java/org/kie/kogito/persistence/jdbc/DDLRunner.java
@@ -15,23 +15,16 @@
  */
 package org.kie.kogito.persistence.jdbc;
 
-import java.util.concurrent.atomic.AtomicBoolean;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 public class DDLRunner {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(DDLRunner.class);
-    private static final AtomicBoolean INITIALIZED = new AtomicBoolean(false);
 
     public static void init(Repository repository, boolean autoDDL) {
         if (!autoDDL) {
             LOGGER.debug("Auto DDL is disabled, do not running initializer scripts");
-            return;
-        }
-        if (INITIALIZED.getAndSet(true)) {
-            LOGGER.debug("DDL runner already initialized");
             return;
         }
         try {
