+++ quarkus/addons/persistence/jdbc/runtime/src/main/java/org/kie/kogito/persistence/quarkus/JDBCorrelationServiceProducer.java
@@ -15,15 +15,32 @@
  */
 package org.kie.kogito.persistence.quarkus;
 
+import java.sql.Connection;
+import java.sql.SQLException;
+
 import javax.enterprise.inject.Produces;
 import javax.sql.DataSource;
 
+import org.kie.kogito.correlation.CorrelationService;
+import org.kie.kogito.persistence.jdbc.DatabaseType;
 import org.kie.kogito.persistence.jdbc.correlation.JDBCCorrelationService;
+import org.kie.kogito.services.event.correlation.DefaultCorrelationService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class JDBCorrelationServiceProducer {
 
+    private static final Logger LOGGER = LoggerFactory.getLogger(JDBCorrelationServiceProducer.class);
+
     @Produces
-    public JDBCCorrelationService jdbcCorrelationService(DataSource dataSource) {
+    public CorrelationService jdbcCorrelationService(DataSource dataSource) {
+        try (Connection connection = dataSource.getConnection()) {
+            if (!DatabaseType.POSTGRES.equals(DatabaseType.getDataBaseType(connection))) {
+                return new DefaultCorrelationService();
+            }
+        } catch (SQLException e) {
+            LOGGER.error("Error getting connection for {}", dataSource);
+        }
         return new JDBCCorrelationService(dataSource);
     }
 }
